made new mechanic of combining primary elements
A circle gets a charge when elements have been destroyed inside it. The
player can clcik on the circle to spend that charge to combine primary
elements (rred, green, blue) to secondary elements (yellor, magenta,
cyan). This works only if there just 2 primary elements that are
compatible. Also circles can now have more the one element of the same
color.